Manual deploy as of commit:6b439762c0432b050238dcd72a570efb87a2f490

+++ smart-contracts/contracts/mixins/MixinLockCore.sol
@@ -114,7 +114,7 @@ contract MixinLockCore is
   // The version number of the current implementation on this network
   function publicLockVersion(
   ) public pure
-    returns (uint)
+    returns (uint16)
   {
     return 9;
   }
